# [compiler] add optimization rules