--- a/compiler/pipes/check-func-calls-and-vararg.cpp
+++ b/compiler/pipes/check-func-calls-and-vararg.cpp
@@ -11,60 +11,6 @@
 #include "compiler/name-gen.h"
 #include "compiler/type-hint.h"
 
-namespace {
-
-bool is_const_bool(VertexPtr expr, bool value) {
-  if (auto as_bool_conv = expr.try_as<op_conv_bool>()) {
-    expr = as_bool_conv->expr();
-  }
-  expr = VertexUtil::get_actual_value(expr);
-  auto target_op = value ? op_true : op_false;
-  return expr->type() == target_op;
-}
-
-VertexAdaptor<op_func_call> replace_call_func(VertexAdaptor<op_func_call> call, const std::string &func_name, std::vector<VertexPtr> args) {
-  auto new_call = VertexAdaptor<op_func_call>::create(std::move(args)).set_location(call);
-  new_call->set_string(func_name);
-  new_call->func_id = G->get_function(func_name);
-  return new_call;
-}
-
-VertexAdaptor<op_func_call> process_microtime(VertexAdaptor<op_func_call> call) {
-  const auto &args = call->args();
-
-  // microtime()      -> microtime_string()
-  // microtime(false) -> microtime_string()
-  if (args.empty() || is_const_bool(args[0], false)) {
-    return replace_call_func(call, "microtime_string", {});
-  }
-
-  // microtime(true) -> microtime_float()
-  if (!args.empty() && is_const_bool(args[0], true)) {
-    return replace_call_func(call, "microtime_float", {});
-  }
-
-  return call;
-}
-
-VertexAdaptor<op_func_call> process_hrtime(VertexAdaptor<op_func_call> call) {
-  const auto &args = call->args();
-
-  // hrtime()      -> hrtime_array()
-  // hrtime(false) -> hrtime_array()
-  if (args.empty() || is_const_bool(args[0], false)) {
-    return replace_call_func(call, "hrtime_array", {});
-  }
-
-  // hrtime(true) -> hrtime_int()
-  if (!args.empty() && is_const_bool(args[0], true)) {
-    return replace_call_func(call, "hrtime_int", {});
-  }
-
-  return call;
-}
-
-} // namespace
-
 /*
  * This pass checks that func calls are correct, i.e. provided necessary amount of arguments, etc.
  * Here we also replace ...variadic call arguments with a single array
@@ -265,12 +211,6 @@ VertexPtr CheckFuncCallsAndVarargPass::create_CompileTimeLocation_call_arg(const
 // (not to carry auto-generated inheritors body, as they are generated wrong anyway, cause JsonEncoder is built-in)
 VertexPtr CheckFuncCallsAndVarargPass::maybe_replace_extern_func_call(VertexAdaptor<op_func_call> call, FunctionPtr f_called) {
   const std::string &f_name = f_called->name;
-  if (f_name == "hrtime") {
-    return process_hrtime(call);
-  }
-  if (f_name == "microtime") {
-    return process_microtime(call);
-  }
   if (f_called->modifiers.is_static() && f_called->class_id) {
     vk::string_view local_name = f_called->local_name();