… builtin-functions/_functions.txt
… builtin-functions/kphp_internal.txt
… compiler/rewrite-rules/impl/rules_generator.py
… compiler/rewrite-rules/impl/rules_lexer.py
… compiler/rewrite-rules/impl/rules_parser.py
… compiler/rewrite-rules/impl/vertex_schema.py
… compiler/rewrite-rules/rules-gen.py
… compiler/rewrite-rules/rules_runtime.cpp
… compiler/rewrite-rules/rules_runtime.h
… compiler/vertex-meta_op_base.h
… compiler/vertex-util.cpp
… compiler/vertex-util.h
… docs/kphp-internals/developing-and-extending-kphp/rewrite-rules.md
… runtime/array.inl
… runtime/array_decl.inl
… runtime/array_functions.cpp
… runtime/array_functions.h
… runtime/conversions_types.inl
… runtime/datetime/datetime_functions.cpp
… runtime/datetime/datetime_functions.h
… runtime/kphp_core.h
… runtime/mixed.inl
… runtime/mixed_decl.inl
… runtime/string.inl
… runtime/string_decl.inl
… runtime/string_functions.cpp
… runtime/string_functions.h
… tests/benchmarks/BenchmarkExplode.php
… tests/benchmarks/BenchmarkStrconv.php
… tests/benchmarks/BenchmarkTmpString.php