Related changes:

* made `hrtime_int` and other specialized function internal
* moved `hrtime` and `microtime` replacement to the rewrite rules
* introduced a `tmp_string` concept
* added a `string::append(tmp_string)` overloading
* added `wrap_substr_args` to re-use substr start+length handling (it's weird)
* added new kphp phpdoc tags for internal use
* restriction-isset learned how to check the optimized functions
* added `explode` specializations that return tuples, etc
* added internal functions aliases if they're replacing existing functions
    * for example, instead of `microtime_string` the user would get `microtime` now

`explode` benchmark results:

    name                old time/op    new time/op    delta
    ExplodeToArray      416ns ± 0%     408ns ± 0%   -1.95%  (p=0.000 n=9+9)
    ExplodeNth          450ns ± 1%     153ns ± 0%  -66.00%  (p=0.000 n=10+7)
    Explode1            457ns ± 1%     143ns ± 1%  -68.78%  (p=0.000 n=10+10)
    Explode2            504ns ± 2%     276ns ± 1%  -45.32%  (p=0.000 n=10+10)
    Explode3            913ns ± 1%     427ns ± 0%  -53.28%  (p=0.000 n=10+8)
    Explode4            951ns ± 1%     564ns ± 0%  -40.69%  (p=0.000 n=10+10)
    Explode3discard     866ns ± 2%     290ns ± 1%  -66.46%  (p=0.000 n=9+10)
    Explode2limit       479ns ± 1%     297ns ± 1%  -38.12%  (p=0.000 n=10+10)
    Explode3limit       835ns ± 2%     471ns ± 0%  -43.54%  (p=0.000 n=8+9)
    Explode4limit       918ns ± 2%     609ns ± 2%  -33.64%  (p=0.000 n=10+10)
    [Geo mean]          642ns          329ns       -48.85%

    name                      old alloc/op   new alloc/op   delta
    ExplodeToArray       344B ± 0%      344B ± 0%        ~  (all equal)
    ExplodeNth           344B ± 0%       40B ± 0%  -88.37%  (p=0.000 n=10+10)
    Explode1             344B ± 0%       40B ± 0%  -88.37%  (p=0.000 n=10+10)
    Explode2             344B ± 0%       80B ± 0%  -76.74%  (p=0.000 n=10+10)
    Explode3             688B ± 0%      112B ± 0%  -83.72%  (p=0.000 n=10+10)
    Explode4             688B ± 0%      160B ± 0%  -76.74%  (p=0.000 n=10+10)
    Explode3discard      688B ± 0%       56B ± 0%  -91.86%  (p=0.000 n=10+10)
    Explode2limit        352B ± 0%       80B ± 0%  -77.27%  (p=0.000 n=10+10)
    Explode3limit        592B ± 0%      112B ± 0%  -81.08%  (p=0.000 n=10+10)
    Explode4limit        624B ± 0%      160B ± 0%  -74.36%  (p=0.000 n=10+10)
    [Geo mean]           476B            96B       -79.92%

    name                 old allocs/op  new allocs/op  delta
    ExplodeToArray       9.00 ± 0%      9.00 ± 0%        ~  (all equal)
    ExplodeNth           9.00 ± 0%      2.00 ± 0%  -77.78%  (p=0.000 n=10+10)
    Explode1             9.00 ± 0%      2.00 ± 0%  -77.78%  (p=0.000 n=10+10)
    Explode2             9.00 ± 0%      4.00 ± 0%  -55.56%  (p=0.000 n=10+10)
    Explode3             17.0 ± 0%       6.0 ± 0%  -64.71%  (p=0.000 n=10+10)
    Explode4             17.0 ± 0%       8.0 ± 0%  -52.94%  (p=0.000 n=10+10)
    Explode3discard      17.0 ± 0%       3.0 ± 0%  -82.35%  (p=0.000 n=10+10)
    Explode2limit        9.00 ± 0%      4.00 ± 0%  -55.56%  (p=0.000 n=10+10)
    Explode3limit        15.0 ± 0%       6.0 ± 0%  -60.00%  (p=0.000 n=10+10)
    Explode4limit        16.0 ± 0%       8.0 ± 0%  -50.00%  (p=0.000 n=10+10)
    [Geo mean]           12.1            4.6       -62.35%

`substr`/`trim` benchmark results:

    name                          old time/op    new time/op    delta
    TmpString::IntvalSubstr          137ns ± 1%      90ns ± 0%   -34.43%  (p=0.000 n=8+10)
    TmpString::AppendSubstr          185ns ± 2%     131ns ± 2%   -29.42%  (p=0.000 n=9+9)
    TmpString::Concat2               149ns ± 1%      91ns ± 1%   -38.94%  (p=0.000 n=10+9)
    TmpString::Concat3               263ns ± 1%     152ns ± 1%   -42.29%  (p=0.000 n=10+8)
    TmpString::ConcatSubstrConv      112ns ± 1%      75ns ± 6%   -33.10%  (p=0.000 n=9+10)
    TmpString::Substr3               179ns ± 0%      91ns ± 0%   -49.08%  (p=0.000 n=10+8)
    TmpString::IndexTrimSubstr       203ns ± 2%     144ns ± 1%   -29.20%  (p=0.000 n=10+10)
    TmpString::IndexSubstr           488ns ± 0%     376ns ± 0%   -22.85%  (p=0.000 n=9+10)
    TmpString::TrimSubstrA           150ns ± 1%     122ns ± 0%   -18.60%  (p=0.000 n=10+8)
    TmpString::TrimSubstrB           152ns ± 1%     121ns ± 1%   -20.60%  (p=0.000 n=10+10)
    TmpString::TrimSubstrC           117ns ± 3%     117ns ± 0%         ~  (p=0.015 n=10+8)
    TmpString::SubstrBoolContext    53.5ns ± 3%    28.0ns ± 0%   -47.66%  (p=0.000 n=10+9)
    TmpString::TrimBoolContext       198ns ± 1%     127ns ± 0%   -35.97%  (p=0.000 n=9+8)
    [Geo mean]                       163ns          110ns        -32.14%

    name                          old alloc/op   new alloc/op   delta
    TmpString::IntvalSubstr          48.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
    TmpString::AppendSubstr           160B ± 0%       96B ± 0%   -40.00%  (p=0.000 n=10+10)
    TmpString::Concat2                104B ± 0%       40B ± 0%   -61.54%  (p=0.000 n=10+10)
    TmpString::Concat3                152B ± 0%       56B ± 0%   -63.16%  (p=0.000 n=10+10)
    TmpString::ConcatSubstrConv       104B ± 0%       56B ± 0%   -46.15%  (p=0.000 n=10+10)
    TmpString::Substr3               80.0B ± 0%     24.0B ± 0%   -70.00%  (p=0.000 n=10+10)
    TmpString::IndexTrimSubstr       56.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
    TmpString::IndexSubstr           64.0B ± 0%     32.0B ± 0%   -50.00%  (p=0.000 n=10+10)
    TmpString::TrimSubstrA           64.0B ± 0%     32.0B ± 0%   -50.00%  (p=0.000 n=10+10)
    TmpString::TrimSubstrB           64.0B ± 0%     32.0B ± 0%   -50.00%  (p=0.000 n=10+10)
    TmpString::TrimSubstrC           32.0B ± 0%     32.0B ± 0%         ~  (all equal)
    TmpString::SubstrBoolContext     24.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
    TmpString::TrimBoolContext       64.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)

    name                          old allocs/op  new allocs/op  delta
    TmpString::IntvalSubstr           2.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
    TmpString::AppendSubstr           4.00 ± 0%      2.00 ± 0%   -50.00%  (p=0.000 n=10+10)
    TmpString::Concat2                3.00 ± 0%      1.00 ± 0%   -66.67%  (p=0.000 n=10+10)
    TmpString::Concat3                4.00 ± 0%      1.00 ± 0%   -75.00%  (p=0.000 n=10+10)
    TmpString::ConcatSubstrConv       2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=10+10)
    TmpString::Substr3                3.00 ± 0%      1.00 ± 0%   -66.67%  (p=0.000 n=10+10)
    TmpString::IndexTrimSubstr        2.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
    TmpString::IndexSubstr            4.00 ± 0%      2.00 ± 0%   -50.00%  (p=0.000 n=10+10)
    TmpString::TrimSubstrA            2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=10+10)
    TmpString::TrimSubstrB            2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=10+10)
    TmpString::TrimSubstrC            1.00 ± 0%      1.00 ± 0%         ~  (all equal)
    TmpString::SubstrBoolContext      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
    TmpString::TrimBoolContext        2.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)

string conversion benchmark results:

    name                                  old time/op    new time/op    delta
    Strconv::ConcatEmptyAndInt32small       44.0ns ± 0%    14.0ns ± 0%   -68.18%  (p=0.002 n=8+10)
    Strconv::ConcatEmptyAndMixed            61.4ns ± 1%    13.0ns ± 0%   -78.83%  (p=0.000 n=10+9)
    Strconv::StringBuildSingleInt32small    43.3ns ± 2%    14.0ns ± 0%   -67.67%  (p=0.000 n=10+10)
    Strconv::StringBuildSingleMixed         59.0ns ± 0%    12.0ns ± 0%   -79.66%  (p=0.000 n=10+10)
    Strconv::ConcatOperand                   115ns ± 1%      72ns ± 1%   -37.02%  (p=0.000 n=10+10)
    [Geo mean]                              60.2ns         23.9ns        -60.29%

    name                                  old alloc/op   new alloc/op   delta
    Strconv::ConcatEmptyAndInt32small        40.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
    Strconv::ConcatEmptyAndMixed             24.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
    Strconv::StringBuildSingleInt32small     40.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
    Strconv::StringBuildSingleMixed          24.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
    Strconv::ConcatOperand                   56.0B ± 0%     40.0B ± 0%   -28.57%  (p=0.000 n=10+10)

    name                                  old allocs/op  new allocs/op  delta
    Strconv::ConcatEmptyAndInt32small         1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
    Strconv::ConcatEmptyAndMixed              1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
    Strconv::StringBuildSingleInt32small      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
    Strconv::StringBuildSingleMixed           1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
    Strconv::ConcatOperand                    2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=10+10)